requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.1.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
cssselect2==0.8.0
Django==5.1.5
django-extensions==3.2.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
fonttools==4.59.0
frozenlist==1.5.0
html5lib==1.1
idna==3.10
kombu==5.4.2
lxml==6.0.0
multidict==6.1.0
mysqlclient==2.2.7
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMySQL==1.1.1
pypdf==5.9.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.4.6
tzdata==2025.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
yarl==1.18.3
zopfli==0.2.3.post1

gunicorn
psycopg2-binary
dj-database-url


